Yodawy quantity hotfix 159

+++ yodawy/src/main/java/com/neqabty/yodawy/modules/orders/domain/entity/OrderItemEntity.kt
@@ -19,7 +19,7 @@ data class OrderItemEntity(
     val price: Double?,
     val priceBeforeSale: Double?,
     val productType: String?,
-    val quantity: Int?,
+    val quantity: Float?,
     val selectedUnit: Int?,
     val variationSKU: String?,
     val wasLimited: Boolean,
